requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
pycryptodome>=3.18.0
cryptography>=41.0.0
scipy>=1.10.0
pydub>=0.25.1
colorama>=0.4.6
tqdm>=4.65.0
